src/server/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.3.8
async-timeout==5.0.1
attrs==24.3.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coverage==7.6.9
cryptography==44.0.0
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pylint==3.3.3
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tomlkit==0.13.2
typing_extensions==4.12.2
upstash-redis==1.2.0
urllib3==2.3.0
uvicorn==0.32.0
yarl==1.18.3
